CORSAIR-178
This was just an unneeded depencency in src/kernel/Makefile. The C++11
standard includes many random number generators, and by default 
a Mersenne Twister 19937 random number generator is used. In the past
the C++ standard libraries did not have anything for generating
Gaussian random numbers, which was the original reason why a Numerical
Recipes was used.